modify bagrmann tests
