+++ tests/test_lab/test_representations/test_bargmanndm.py
@@ -15,9 +15,13 @@
 import numpy as np
 import pytest
 
-from mrmustard.lab.representations.bargmann_ket import BargmannKet
+from mrmustard.lab.representations.bargmann_dm import BargmannDM
 
 
-def test_purity_of_bargmann_ket():
-        bargmannket = BargmannKet(A=np.random.random((3,3)), b=np.random.random(3), c=1.0)
-        assert(bargmannket.purity, 1.0)
+class TestBargmannKetThrowErrors():
+
+    bargmanndm = BargmannDM(A=np.random.random((3,3)), b=np.random.random(3), c=1.0)
+
+    def test_purity_with_error(self):
+        with self.assertRaises(NotImplementedError):
+            self.bargmanndm.purity
